+++ CHANGELOG.md
@@ -124,6 +124,7 @@
 * Fix the broken search results table. Fixes [UILD-788].
 * Importing an empty file treated as an error. Fixes [UILD-692].
 * Fix CSS rule leaking out of this module. Fixes [UILD-796].
+* Fix overlapping labels in the search results table. Fixes [UILD-797].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -245,6 +246,7 @@
 [UILD-788]:https://folio-org.atlassian.net/browse/UILD-788
 [UILD-692]:https://folio-org.atlassian.net/browse/UILD-692
 [UILD-796]:https://folio-org.atlassian.net/browse/UILD-796
+[UILD-797]:https://folio-org.atlassian.net/browse/UILD-797
 
 ## 1.0.5 (2025-04-30)
 * Browser Back button navigates to Edit without duplicated title when navigated from Duplicate screen. Fixes [UILD-554].