+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Change history for ui-linked-data
 
 ## 3.0.0 (IN PROGRESS)
+
+## 2.0.1 (2026-04-23)
 * Fix for the case when changes to profile settings are not applied immediately. Fixes [UILD-798].
+* Fix overlapping labels in the search results table. Fixes [UILD-797].
 
 [UILD-798]:https://folio-org.atlassian.net/browse/UILD-798
+[UILD-797]:https://folio-org.atlassian.net/browse/UILD-797
 
 ## 2.0.0 (2026-04-17)
 * Remove unnecessary child subcomponents when copying a whole field when the record is loaded. Fixes [UILD-544].
@@ -127,7 +131,6 @@
 * Fix the broken search results table. Fixes [UILD-788].
 * Importing an empty file treated as an error. Fixes [UILD-692].
 * Fix CSS rule leaking out of this module. Fixes [UILD-796].
-* Fix overlapping labels in the search results table. Fixes [UILD-797].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -249,7 +252,6 @@
 [UILD-788]:https://folio-org.atlassian.net/browse/UILD-788
 [UILD-692]:https://folio-org.atlassian.net/browse/UILD-692
 [UILD-796]:https://folio-org.atlassian.net/browse/UILD-796
-[UILD-797]:https://folio-org.atlassian.net/browse/UILD-797
 
 ## 1.0.5 (2025-04-30)
 * Browser Back button navigates to Edit without duplicated title when navigated from Duplicate screen. Fixes [UILD-554].